+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/actions/gc/TriggerGcCollectionAction.kt
@@ -1,0 +1,26 @@
+package com.github.rafaelldi.diagnosticsclientplugin.actions.gc
+
+import com.github.rafaelldi.diagnosticsclientplugin.generated.diagnosticsHostModel
+import com.github.rafaelldi.diagnosticsclientplugin.services.TriggerGcCollectionController
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.components.service
+import com.jetbrains.rider.projectView.solution
+
+class TriggerGcCollectionAction : AnAction() {
+    override fun actionPerformed(event: AnActionEvent) {
+        val project = event.project ?: return
+        val pid = project.solution.diagnosticsHostModel.processList.selected.value ?: return
+        val controller = project.service<TriggerGcCollectionController>()
+        controller.triggerGc(pid)
+    }
+
+    override fun update(event: AnActionEvent) {
+        val project = event.project
+        if (project == null) {
+            event.presentation.isEnabled = false
+        } else {
+            event.presentation.isEnabled = project.solution.diagnosticsHostModel.processList.selected.value != null
+        }
+    }
+}